Revert moving base class, would require moving properties too
