+++ user/test/com/google/gwt/i18n/client/DateTimeFormat_en_Test.java
@@ -18,6 +18,7 @@ package com.google.gwt.i18n.client;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.i18n.client.constants.TimeZoneConstants;
 import com.google.gwt.i18n.client.impl.cldr.DateTimeFormatInfoImpl_de;
+import com.google.gwt.i18n.shared.DateTimeFormatTestBase;
 
 import java.util.Date;
