+++ user/test/com/google/gwt/i18n/shared/DateTimeFormatTestBase.java
@@ -13,9 +13,9 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.gwt.i18n.client;
+package com.google.gwt.i18n.shared;
 
-import com.google.gwt.i18n.shared.DateTimeFormat;
+import com.google.gwt.i18n.client.Messages;
 import com.google.gwt.i18n.shared.DateTimeFormat.PredefinedFormat;
 import com.google.gwt.junit.client.GWTTestCase;
